moved BagitWarning to the same directory as the linter as that makes more sense
